requirements.txt
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
pydrive2

pika
httpx
anyio
requests

python-dotenv

sentry_sdk

openai-whisper
setuptools-rust
tensorboard
openai

fuzzywuzzy
python-Levenshtein